+++ sdk/cosmos/azure-cosmos/tests/perfstress_tests/full_text_search_perf.py
@@ -18,8 +18,8 @@ class FullTextSearchTest(PerfStressTest):
         self.logging = False
         self.latency = False
         self.ru = False
-        self.queryIndex = 1    # 0 for full text search, 1 for full text rank, and 2 for hybrid search
-        top = 1001
+        self.queryIndex = 2    # 0 for full text search, 1 for full text rank, and 2 for hybrid search
+        top = 1000
 
         self.client = CosmosClient(TestConfig.host, credential=TestConfig.credential)
         if (self.logging):
@@ -33,12 +33,12 @@ class FullTextSearchTest(PerfStressTest):
             logger.addHandler(handler)
             self.client = CosmosClient(TestConfig.host, credential=TestConfig.credential, logger=logger, enable_diagnostics_logging=True)
 
-        database = self.client.get_database_client('QueryHybridRankTesting')
-        self.container = database.get_container_client('arxiv-250kdocuments-index')
+        database = self.client.get_database_client('perf-tests-sdks')
+        self.container = database.get_container_client('fts')
         embedding = [random.uniform(-1, 1) for _ in range(128)]
         self.queries = ["SELECT TOP " + str(top) + " c.id AS Text FROM c WHERE FullTextContains(c.abstract, 'shoulder')",
-                        "SELECT TOP " + str(top) + " c.id AS Text FROM c Order By Rank FullTextScore(c.abstract, ['may', 'music'])",
-                        "SELECT TOP " + str(top) + " c.text AS text FROM c ORDER BY RANK RRF(FullTextScore(c.text, ['may', 'music']), VectorDistance(c.vector," + str(embedding) + ")) "]
+                        "SELECT TOP " + str(top) + " c.id AS Text FROM c Order By Rank FullTextScore(c.text, ['may', 'music'])",
+                        "SELECT TOP " + str(top) + " c.id AS text FROM c ORDER BY RANK RRF(FullTextScore(c.abstract, ['may', 'music']), VectorDistance(c.vector," + str(embedding) + ")) "]
         self.top_stats = []
         self.h = hpy()
         """ self.tracer = VizTracer()