+++ sdk/cosmos/azure-cosmos/tests/test_full_text_policy.py
@@ -71,7 +71,17 @@ class TestFullTextPolicy(unittest.TestCase):
         self.test_db.delete_container(created_container.id)
 
     def test_replace_full_text_container(self):
+<<<<<<< HEAD:sdk/cosmos/azure-cosmos/tests/test_full_text_policy.py
         # Replace a container with a valid full text policy and full text indexing policy
+=======
+        # Replace a container without a full text policy and full text indexing policy
+
+        created_container = self.test_db.create_container(
+            id='full_text_container' + str(uuid.uuid4()),
+            partition_key=PartitionKey(path="/id")
+        )
+        created_container_properties = created_container.read()
+>>>>>>> 1fb4dfbc37e58a39e0fbd911b0df091986401e52:sdk/cosmos/azure-cosmos/test/test_full_text_policy.py
         full_text_policy = {
             "defaultLanguage": "en-US",
             "fullTextPaths": [
@@ -86,6 +96,7 @@ class TestFullTextPolicy(unittest.TestCase):
                 {"path": "/abstract"}
             ]
         }
+<<<<<<< HEAD:sdk/cosmos/azure-cosmos/tests/test_full_text_policy.py
         created_container = self.test_db.create_container(
             id='full_text_container' + str(uuid.uuid4()),
             partition_key=PartitionKey(path="/id"),
@@ -101,6 +112,12 @@ class TestFullTextPolicy(unittest.TestCase):
         indexing_policy['fullTextIndexes'][0]['path'] = "/new_path"
         replaced_container = self.test_db.create_container(
             id=created_container.id,
+=======
+
+        # Replace the container with new policies
+        replaced_container = self.test_db.replace_container(
+            container=created_container.id,
+>>>>>>> 1fb4dfbc37e58a39e0fbd911b0df091986401e52:sdk/cosmos/azure-cosmos/test/test_full_text_policy.py
             partition_key=PartitionKey(path="/id"),
             full_text_policy=full_text_policy,
             indexing_policy=indexing_policy
@@ -108,6 +125,37 @@ class TestFullTextPolicy(unittest.TestCase):
         properties = replaced_container.read()
         assert properties["fullTextPolicy"] == full_text_policy
         assert properties["indexingPolicy"]['fullTextIndexes'] == indexing_policy['fullTextIndexes']
+<<<<<<< HEAD:sdk/cosmos/azure-cosmos/tests/test_full_text_policy.py
+=======
+        assert created_container_properties['indexingPolicy'] != properties['indexingPolicy']
+        self.test_db.delete_container(created_container.id)
+
+        # Replace a container with a valid full text policy and full text indexing policy
+        created_container = self.test_db.create_container(
+            id='full_text_container' + str(uuid.uuid4()),
+            partition_key=PartitionKey(path="/id"),
+            full_text_policy=full_text_policy,
+            indexing_policy=indexing_policy
+        )
+        created_container_properties = created_container.read()
+        assert properties["fullTextPolicy"] == full_text_policy
+        assert properties["indexingPolicy"]['fullTextIndexes'] == indexing_policy['fullTextIndexes']
+
+        # Replace the container with new policies
+        full_text_policy['fullTextPaths'][0]['path'] = "/new_path"
+        indexing_policy['fullTextIndexes'][0]['path'] = "/new_path"
+        replaced_container = self.test_db.replace_container(
+            container=created_container.id,
+            partition_key=PartitionKey(path="/id"),
+            full_text_policy=full_text_policy,
+            indexing_policy=indexing_policy
+        )
+        properties = replaced_container.read()
+        assert properties["fullTextPolicy"] == full_text_policy
+        assert properties["indexingPolicy"]['fullTextIndexes'] == indexing_policy['fullTextIndexes']
+        assert created_container_properties['fullTextPolicy'] != properties['fullTextPolicy']
+        assert created_container_properties["indexingPolicy"] != properties["indexingPolicy"]
+>>>>>>> 1fb4dfbc37e58a39e0fbd911b0df091986401e52:sdk/cosmos/azure-cosmos/test/test_full_text_policy.py
         self.test_db.delete_container(created_container.id)
 
     def test_fail_create_full_text_policy(self):