+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 #### Features Added
 * Added full text policy and full text indexing policy. See [PR 37891](https://github.com/Azure/azure-sdk-for-python/pull/37891).
+<<<<<<< HEAD
+=======
+* Added support for full text search and hybrid search queries. See [PR 38275](https://github.com/Azure/azure-sdk-for-python/pull/38275).
+>>>>>>> 1fb4dfbc37e58a39e0fbd911b0df091986401e52
 
 #### Breaking Changes